Se agrega repository de Rubra

+++ src/main/java/com/example/springsocial/security/TokenAuthenticationFilter.java
@@ -33,8 +33,8 @@ public class TokenAuthenticationFilter extends OncePerRequestFilter {
 
             if (StringUtils.hasText(jwt) && tokenProvider.validateToken(jwt)) {
                 Long userId = tokenProvider.getUserIdFromToken(jwt);
-
-                UserDetails userDetails = customUserDetailsService.loadUserById(userId);
+                //TODO revisar aca si esta bien
+                UserDetails userDetails = customUserDetailsService.loadUserById(userId.intValue());
                 UsernamePasswordAuthenticationToken authentication = new UsernamePasswordAuthenticationToken(userDetails, null, userDetails.getAuthorities());
                 authentication.setDetails(new WebAuthenticationDetailsSource().buildDetails(request));
